+++ frontend/src/components/pages/Browse.js
@@ -6,47 +6,13 @@ import * as listingApi from "../../utils/api/listing-api";
 import * as userApi from "../../utils/api/user-api";
 import update from "immutability-helper";
 
-const Browse = () => {
-  const [user, setUser] = useState(JSON.parse(localStorage.getItem("user")));
-
+const Browse = ({ user, setUser, addFavorite, removeFavorite }) => {
   const [listings, setListings] = useState([]);
 
   const getListings = async () => {
     const response = await listingApi.searchListings(user.token, {});
     const responseJson = await response.json();
     setListings(responseJson);
-  };
-
-  const addFavorite = async (listingId) => {
-    try {
-      const response = await userApi.addFavorite(user.token, listingId);
-      const responseJson = await response.json();
-      console.log("REST API response", responseJson);
-      if (response.ok) {
-        const updatedUser = update(user, {
-          favoriteListings: { $set: responseJson.favoriteListings },
-        });
-        setUser(updatedUser);
-      }
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  const removeFavorite = async (listingId) => {
-    try {
-      const response = await userApi.removeFavorite(user.token, listingId);
-      const responseJson = await response.json();
-      console.log("REST API response", responseJson);
-      if (response.ok) {
-        const updatedUser = update(user, {
-          favoriteListings: { $set: responseJson.favoriteListings },
-        });
-        setUser(updatedUser);
-      }
-    } catch (err) {
-      console.log(err);
-    }
   };
 
   const isListingFavorited = (listingId) => {
@@ -72,12 +38,7 @@ const Browse = () => {
 
   useEffect(() => {
     getListings();
-  }, [user]);
-
-  // Update localStorage every time user info gets updated (e.g. if they create/delete a listing)
-  useEffect(() => {
-    localStorage.setItem("user", JSON.stringify(user));
-  }, [user]);
+  });
 
   return (
     <>
